Fix metadata on remove actions
When an element is removed from a relation (remove action), in adiff format,
the metadata of the action is the one of the latest version of the removed
element.

We fix it so that the metadata is the metadata of the remove operation itself.

Introduce the notion of "patcher", which modifies the data

ref #12 @3h30